refactors a bunch of others
