+++ hdk-rust/src/api/debug.rs
